Revert "PLT-635 Trigger policy cache refresh job on demand"
