+++ auth-agents-common/src/main/java/org/apache/atlas/ranger/plugin/util/DownloaderTask.java
@@ -35,14 +35,6 @@ public final class  DownloaderTask extends TimerTask {
         this.queue = queue;
     }
 
-    public DownloaderTask(BlockingQueue<DownloadTrigger> queue,
-                          boolean policies, boolean roles, boolean groups) {
-        this.queue = queue;
-        timerTrigger.policies = policies;
-        timerTrigger.roles = roles;
-        timerTrigger.groups = groups;
-    }
-
     @Override
     public void run() {
         try {
