+++ auth-plugin-atlas/src/main/java/org/apache/atlas/ranger/authorization/atlas/authorizer/RangerAtlasAuthorizer.java
@@ -141,11 +141,6 @@ public class RangerAtlasAuthorizer implements AtlasAuthorizer {
     }
 
     @Override
-    public void refreshCache(boolean policies, boolean roles, boolean groups) {
-        atlasPlugin.submitRefresherTask(policies, roles, groups);
-    }
-
-    @Override
     public void cleanUp() {
         if (LOG.isDebugEnabled()) {
             LOG.debug("==> cleanUp ");
