+++ authorization/src/main/java/org/apache/atlas/authorize/AtlasAuthorizationUtils.java
@@ -270,19 +270,6 @@ public class AtlasAuthorizationUtils {
         return ret;
     }
 
-    public static void refreshCache(boolean policies, boolean roles, boolean groups) {
-        try {
-            AtlasAuthorizer authorizer = AtlasAuthorizerFactory.getAtlasAuthorizer();
-            if (authorizer == null ) {
-                throw new AtlasAuthorizationException("Authorizer is null");
-            }
-
-            authorizer.refreshCache(policies, roles, groups);
-        } catch (AtlasAuthorizationException e) {
-            LOG.error("Unable to obtain AtlasAuthorizer", e);
-        }
-    }
-
     public static void filterTypesDef(AtlasTypesDefFilterRequest request) {
         MetricRecorder metric  = RequestContext.get().startMetricRecord("filterTypesDef");
         String        userName = getCurrentUserName();
