+++ common/src/main/java/org/apache/atlas/repository/Constants.java
@@ -141,7 +141,7 @@ public final class Constants {
     public static final String QUERY_ENTITY_TYPE            = "Query";
     public static final String README_ENTITY_TYPE           = "Readme";
     public static final String QUERY_FOLDER_ENTITY_TYPE     = "Folder";
-    public static final String COLLECTION_ENTITY_TYPE       = "Collection";
+    public static final String QUERY_COLLECTION_ENTITY_TYPE = "Collection";
 
     public static final String ACCESS_CONTROL_RELATION_TYPE  = "access_control_policies";
