+++ repository/src/main/java/org/apache/atlas/discovery/EntityDiscoveryService.java
@@ -1024,6 +1024,7 @@ public class EntityDiscoveryService implements AtlasDiscoveryService {
             }
 
             String indexName = getIndexName(params);
+            LOG.info("directIndexSearch.indexName {}", indexName);
 
             indexQuery = graph.elasticsearchQuery(indexName);
