+++ webapp/src/main/java/org/apache/atlas/web/rest/AuthREST.java
@@ -19,7 +19,6 @@ package org.apache.atlas.web.rest;
 
 import org.apache.atlas.RequestContext;
 import org.apache.atlas.annotation.Timed;
-import org.apache.atlas.authorize.AtlasAuthorizationUtils;
 import org.apache.atlas.exception.AtlasBaseException;
 import org.apache.atlas.model.audit.AuditSearchParams;
 import org.apache.atlas.model.audit.EntityAuditSearchResult;
@@ -46,7 +45,6 @@ import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
-import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -162,46 +160,6 @@ public class AuthREST {
             updateLastSync(serviceName);
 
             return ret;
-        } finally {
-            AtlasPerfTracer.log(perf);
-        }
-    }
-
-    @POST
-    @Path("refreshcache")
-    @Timed
-    public void refreshCache(@QueryParam("policies") Boolean policies,
-                                        @QueryParam("roles") Boolean roles,
-                                        @QueryParam("groups") Boolean groups) {
-        AtlasPerfTracer perf = null;
-
-        try {
-            if (AtlasPerfTracer.isPerfTraceEnabled(PERF_LOG)) {
-                perf = AtlasPerfTracer.getPerfTracer(PERF_LOG, "AuthREST.refreshCache(policies="+policies+", roles="+roles+", groups="+groups+")");
-            }
-
-            if (policies == null && roles == null && groups == null) {
-                AtlasAuthorizationUtils.refreshCache(true, true, true);
-
-            } else {
-                boolean refreshPolicies = false;
-                boolean refreshRoles = false;
-                boolean refreshGroups = false;
-
-                if (Boolean.TRUE == policies) {
-                    refreshPolicies = true;
-                }
-
-                if (Boolean.TRUE == roles) {
-                    refreshRoles = true;
-                }
-
-                if (Boolean.TRUE == groups) {
-                    refreshGroups = true;
-                }
-
-                AtlasAuthorizationUtils.refreshCache(refreshPolicies, refreshRoles, refreshGroups);
-            }
         } finally {
             AtlasPerfTracer.log(perf);
         }
